RESPONDER: New interface for client registration
This is just a beginning of new responder interface to data provider
and it is just to make the client registration work. It needs further
improvement.

The idea is to take the existing interface and make it work better
with further extensions of data provider. The current interface has
several disadvantages such as it is originally build only for
account requests and doesn't take different set of output parameters.
It also doesn't work well with integration into tevent-made responders.